// src/types/line/line_types.ts
type Int64 = number | bigint;
// ...
export type ContactStatus =
	| 0
	| "UNSPECIFIED"
	| 1
	| "FRIEND"
	| 2
	| "FRIEND_BLOCKED"
	| 3
	| "RECOMMEND"
	| 4
	| "RECOMMEND_BLOCKED"
	| 5
	| "DELETED"
	| 6
	| "DELETED_BLOCKED";

export type ContactType =
	| 0
	| "MID"
	| 1
	| "PHONE"
	| 2
	| "EMAIL"
	| 3
	| "USERID"
	| 4
	| "PROXIMITY"
	| 5
	| "GROUP"
	| 6
	| "USER"
	| 7
	| "QRCODE"
	| 8
	| "PROMOTION_BOT"
	| 9
	| "CONTACT_MESSAGE"
	| 10
	| "FRIEND_REQUEST"
	| 11
	| "BEACON"
	| 128
	| "REPAIR"
	| 2305
	| "FACEBOOK"
	| 2306
	| "SINA"
	| 2307
	| "RENREN"
	| 2308
	| "FEIXIN"
	| 2309
	| "BBM";

export type ContentType =
	| 0
	| "NONE"
	| 1
	| "IMAGE"
	| 2
	| "VIDEO"
	| 3
	| "AUDIO"
	| 4
	| "HTML"
	| 5
	| "PDF"
	| 6
	| "CALL"
	| 7
	| "STICKER"
	| 8
	| "PRESENCE"
	| 9
	| "GIFT"
	| 10
	| "GROUPBOARD"
	| 11
	| "APPLINK"
	| 12
	| "LINK"
	| 13
	| "CONTACT"
	| 14
	| "FILE"
	| 15
	| "LOCATION"
	| 16
	| "POSTNOTIFICATION"
	| 17
	| "RICH"
	| 18
	| "CHATEVENT"
	| 19
	| "MUSIC"
	| 20
	| "PAYMENT"
	| 21
	| "EXTIMAGE"
	| 22
	| "FLEX";
// ...
export type FriendRequestStatus =
	| 0
	| "NONE"
	| 1
	| "AVAILABLE"
	| 2
	| "ALREADY_REQUESTED"
	| 3
	| "UNAVAILABLE";

export type IdentityProvider =
	| 0
	| "UNKNOWN"
	| 1
	| "LINE"
	| 2
	| "NAVER_KR"
	| 3
	| "LINE_PHONE";

export type LN0_F0 =
	| 0
	| "UNKNOWN"
	| 1
	| "INVALID_TARGET_USER"
	| 2
	| "AGE_VALIDATION"
	| 3
	| "TOO_MANY_FRIENDS"
	| 4
	| "TOO_MANY_REQUESTS"
	| 5
	| "MALFORMED_REQUEST"
	| 6
	| "TRACKING_META_QRCODE_FAVORED";

export type LN0_X0 = 1 | "USER" | 2 | "BOT";

export type MIDType =
	| 0
	| "USER"
	| 1
	| "ROOM"
	| 2
	| "GROUP"
	| 3
	| "SQUARE"
	| 4
	| "SQUARE_CHAT"
	| 5
	| "SQUARE_MEMBER"
	| 6
	| "BOT"
	| 7
	| "SQUARE_THREAD";
// ...
export type Pb1_B = 1 | "SIRI" | 2 | "GOOGLE_ASSISTANT" | 3 | "OS_SHARE";

export type Pb1_D0 =
	| 0
	| "RICH_MENU_ID"
	| 1
	| "STATUS_BAR"
	| 2
	| "BUDDY_CAUTION_NOTICE";

export type Pb1_D4 = 1 | "AUDIO" | 2 | "VIDEO" | 3 | "FACEPLAY";

export type Pb1_D6 =
	| 0
	| "GOOGLE"
	| 1
	| "BAIDU"
	| 2
	| "FOURSQUARE"
	| 3
	| "YAHOOJAPAN"
	| 4
	| "KINGWAY";

export type Pb1_E7 = 0 | "UNKNOWN" | 1 | "TALK" | 2 | "SQUARE";
// ...
export type Pb1_EnumC13015h6 =
	| 0
	| "FORWARD"
	| 1
	| "AUTO_REPLY"
	| 2
	| "SUBORDINATE"
	| 3
	| "REPLY";
// ...
export type Pb1_EnumC13151r3 = 0 | "ONEWAY" | 1 | "BOTH" | 2 | "NOT_REGISTERED";
// ...
export type Pb1_N6 =
	| 1
	| "NFT"
	| 2
	| "AVATAR"
	| 3
	| "SNOW"
	| 4
	| "ARCZ"
	| 5
	| "FRENZ";

export type Pb1_O2 =
	| 1
	| "NAME"
	| 2
	| "PICTURE_STATUS"
	| 4
	| "PREVENTED_JOIN_BY_TICKET"
	| 8
	| "NOTIFICATION_SETTING"
	| 16
	| "INVITATION_TICKET"
	| 32
	| "FAVORITE_TIMESTAMP"
	| 64
	| "CHAT_TYPE"
	| 128
	| "PENALTY";

export type Pb1_O6 = 1 | "DEFAULT" | 2 | "MULTI_PROFILE";
// ...
export type Pb1_Z2 = 0 | "GROUP" | 1 | "ROOM" | 2 | "PEER";
// ...
export type SquareEventStatus = 1 | "NORMAL" | 2 | "ALERT_DISABLED";

export type SquareEventType =
	| 0
	| "RECEIVE_MESSAGE"
	| 1
	| "SEND_MESSAGE"
	| 2
	| "NOTIFIED_JOIN_SQUARE_CHAT"
	| 3
	| "NOTIFIED_INVITE_INTO_SQUARE_CHAT"
	| 4
	| "NOTIFIED_LEAVE_SQUARE_CHAT"
	| 5
	| "NOTIFIED_DESTROY_MESSAGE"
	| 6
	| "NOTIFIED_MARK_AS_READ"
	| 7
	| "NOTIFIED_UPDATE_SQUARE_MEMBER_PROFILE"
	| 8
	| "NOTIFIED_UPDATE_SQUARE"
	| 9
	| "NOTIFIED_UPDATE_SQUARE_STATUS"
	| 10
	| "NOTIFIED_UPDATE_SQUARE_AUTHORITY"
	| 11
	| "NOTIFIED_UPDATE_SQUARE_MEMBER"
	| 12
	| "NOTIFIED_UPDATE_SQUARE_CHAT"
	| 13
	| "NOTIFIED_UPDATE_SQUARE_CHAT_STATUS"
	| 14
	| "NOTIFIED_UPDATE_SQUARE_CHAT_MEMBER"
	| 15
	| "NOTIFIED_CREATE_SQUARE_MEMBER"
	| 16
	| "NOTIFIED_CREATE_SQUARE_CHAT_MEMBER"
	| 17
	| "NOTIFIED_UPDATE_SQUARE_MEMBER_RELATION"
	| 18
	| "NOTIFIED_SHUTDOWN_SQUARE"
	| 19
	| "NOTIFIED_KICKOUT_FROM_SQUARE"
	| 20
	| "NOTIFIED_DELETE_SQUARE_CHAT"
	| 21
	| "NOTIFICATION_JOIN_REQUEST"
	| 22
	| "NOTIFICATION_JOINED"
	| 23
	| "NOTIFICATION_PROMOTED_COADMIN"
	| 24
	| "NOTIFICATION_PROMOTED_ADMIN"
	| 25
	| "NOTIFICATION_DEMOTED_MEMBER"
	| 26
	| "NOTIFICATION_KICKED_OUT"
	| 27
	| "NOTIFICATION_SQUARE_DELETE"
	| 28
	| "NOTIFICATION_SQUARE_CHAT_DELETE"
	| 29
	| "NOTIFICATION_MESSAGE"
	| 30
	| "NOTIFIED_UPDATE_SQUARE_CHAT_PROFILE_NAME"
	| 31
	| "NOTIFIED_UPDATE_SQUARE_CHAT_PROFILE_IMAGE"
	| 32
	| "NOTIFIED_UPDATE_SQUARE_FEATURE_SET"
	| 33
	| "NOTIFIED_ADD_BOT"
	| 34
	| "NOTIFIED_REMOVE_BOT"
	| 36
	| "NOTIFIED_UPDATE_SQUARE_NOTE_STATUS"
	| 37
	| "NOTIFIED_UPDATE_SQUARE_CHAT_ANNOUNCEMENT"
	| 38
	| "NOTIFIED_UPDATE_SQUARE_CHAT_MAX_MEMBER_COUNT"
	| 39
	| "NOTIFICATION_POST_ANNOUNCEMENT"
	| 40
	| "NOTIFICATION_POST"
	| 41
	| "MUTATE_MESSAGE"
	| 42
	| "NOTIFICATION_NEW_CHAT_MEMBER"
	| 43
	| "NOTIFIED_UPDATE_READONLY_CHAT"
	| 46
	| "NOTIFIED_UPDATE_MESSAGE_STATUS"
	| 47
	| "NOTIFICATION_MESSAGE_REACTION"
	| 48
	| "NOTIFIED_CHAT_POPUP"
	| 49
	| "NOTIFIED_SYSTEM_MESSAGE"
	| 50
	| "NOTIFIED_UPDATE_SQUARE_CHAT_FEATURE_SET"
	| 51
	| "NOTIFIED_UPDATE_LIVE_TALK"
	| 52
	| "NOTIFICATION_LIVE_TALK"
	| 53
	| "NOTIFIED_UPDATE_LIVE_TALK_INFO"
	| 54
	| "NOTIFICATION_THREAD_MESSAGE"
	| 55
	| "NOTIFICATION_THREAD_MESSAGE_REACTION"
	| 56
	| "NOTIFIED_UPDATE_THREAD"
	| 57
	| "NOTIFIED_UPDATE_THREAD_STATUS"
	| 58
	| "NOTIFIED_UPDATE_THREAD_MEMBER"
	| 59
	| "NOTIFIED_UPDATE_THREAD_ROOT_MESSAGE"
	| 60
	| "NOTIFIED_UPDATE_THREAD_ROOT_MESSAGE_STATUS"
	| 61
	| "NOTIFIED_CREATE_SQUARE_SUBSCRIPTION"
	| 62
	| "NOTIFIED_UPDATE_SQUARE_SUBSCRIPTION";
// ...
export interface AvatarProfile {
	version: string;
	updatedMillis: Int64;
	thumbnail: string;
	usablePublicly: boolean;
}
// ...
export interface Chat {
	type: Pb1_Z2;
	chatMid: string;
	createdTime: Int64;
	notificationDisabled: boolean;
	favoriteTimestamp: Int64;
	chatName: string;
	picturePath: string;
	extra: Pb1_C13208v4;
}
// ...
export interface Contact {
	mid: string;
	createdTime: Int64;
	type: ContactType;
	status: ContactStatus;
	relation: Pb1_EnumC13151r3;
	displayName: string;
	phoneticName: string;
	pictureStatus: string;
	thumbnailUrl: string;
	statusMessage: string;
	displayNameOverridden: string;
	favoriteTime: Int64;
	capableVoiceCall: boolean;
	capableVideoCall: boolean;
	capableMyhome: boolean;
	capableBuddy: boolean;
	attributes: number;
	settings: Int64;
	picturePath: string;
	recommendParams: string;
	friendRequestStatus: FriendRequestStatus;
	musicProfile: string;
	videoProfile: string;
	statusMessageContentMetadata: Record<string, string>;
	avatarProfile: AvatarProfile;
	friendRingtone: string;
	friendRingbackTone: string;
	nftProfile: boolean;
	pictureSource: Pb1_N6;
	profileId: string;
	premiumBlocked: boolean;
	hiddenFromList: boolean;
}
// ...
export interface Location {
	title: string;
	address: string;
	latitude: number;
	longitude: number;
	phone: string;
	categoryId: string;
	provider: Pb1_D6;
	accuracy: GeolocationAccuracy;
	altitudeMeters: number;
}
// ...
export interface Message {
	from: string;
	to: string;
	toType: MIDType;
	id: string;
	createdTime: Int64;
	deliveredTime: Int64;
	text: string;
	location: Location;
	hasContent: boolean;
	contentType: ContentType;
	contentPreview: string;
	contentMetadata: Record<string, string>;
	sessionId: number;
	chunks: (string | Buffer)[];
	relatedMessageId: string;
	messageRelationType: Pb1_EnumC13015h6;
	readCount: number;
	relatedMessageServiceCode: Pb1_E7;
	appExtensionType: Pb1_B;
	reactions: Reaction[];
	updatedTime: Int64;
}
// ...
export interface Pb1_C13208v4 {
	groupExtra: GroupExtra;
	peerExtra: Pb1_A6;
}
// ...
export interface Profile {
	mid: string;
	userid: string;
	phone: string;
	email: string;
	regionCode: string;
	displayName: string;
	phoneticName: string;
	pictureStatus: string;
	thumbnailUrl: string;
	statusMessage: string;
	allowSearchByUserid: boolean;
	allowSearchByEmail: boolean;
	picturePath: string;
	musicProfile: string;
	videoProfile: string;
	statusMessageContentMetadata: Record<string, string>;
	avatarProfile: AvatarProfile;
	nftProfile: boolean;
	pictureSource: Pb1_N6;
	profileId: string;
	profileType: Pb1_O6;
	createdTimeMillis: Int64;
}
// ...
export interface Reaction {
	fromUserMid: string;
	atMillis: Int64;
	reactionType: ReactionType;
}
// ...
export interface SquareEvent {
	createdTime: Int64;
	type: SquareEventType;
	payload: SquareEventPayload;
	syncToken: string;
	eventStatus: SquareEventStatus;
}
// ...
export interface SquareEventNotifiedDestroyMessage {
	squareChatMid: string;
	messageId: string;
	threadMid: string;
}

export interface SquareEventNotifiedInviteIntoSquareChat {
	squareChatMid: string;
	invitees: SquareMember[];
	invitor: SquareMember;
	invitorRelation: SquareMemberRelation;
}

export interface SquareEventNotifiedJoinSquareChat {
	squareChatMid: string;
	joinedMember: SquareMember;
}
// ...
export interface SquareEventReceiveMessage {
	squareChatMid: string;
	squareMessage: SquareMessage;
	senderDisplayName: string;
	messageReactionStatus: SquareMessageReactionStatus;
	senderRevision: Int64;
	squareMid: string;
	threadMid: string;
	threadTotalMessageCount: Int64;
	threadLastMessageAt: Int64;
	contentsAttribute: ContentsAttribute;
}

export interface SquareEventSendMessage {
	squareChatMid: string;
	squareMessage: SquareMessage;
	reqSeq: number;
	senderDisplayName: string;
	messageReactionStatus: SquareMessageReactionStatus;
	threadMid: string;
	threadTotalMessageCount: Int64;
	threadLastMessageAt: Int64;
}
// ...
export interface SquareEventPayload {
	receiveMessage: SquareEventReceiveMessage;
	sendMessage: SquareEventSendMessage;
	notifiedJoinSquareChat: SquareEventNotifiedJoinSquareChat;
	notifiedInviteIntoSquareChat: SquareEventNotifiedInviteIntoSquareChat;
	notifiedLeaveSquareChat: SquareEventNotifiedLeaveSquareChat;
	notifiedDestroyMessage: SquareEventNotifiedDestroyMessage;
	notifiedMarkAsRead: SquareEventNotifiedMarkAsRead;
	notifiedUpdateSquareMemberProfile:
		SquareEventNotifiedUpdateSquareMemberProfile;
	notifiedUpdateSquare: SquareEventNotifiedUpdateSquare;
	notifiedUpdateSquareMember: SquareEventNotifiedUpdateSquareMember;
	notifiedUpdateSquareChat: SquareEventNotifiedUpdateSquareChat;
	notifiedUpdateSquareChatMember: SquareEventNotifiedUpdateSquareChatMember;
	notifiedUpdateSquareAuthority: SquareEventNotifiedUpdateSquareAuthority;
	notifiedUpdateSquareStatus: SquareEventNotifiedUpdateSquareStatus;
	notifiedUpdateSquareChatStatus: SquareEventNotifiedUpdateSquareChatStatus;
	notifiedCreateSquareMember: SquareEventNotifiedCreateSquareMember;
	notifiedCreateSquareChatMember: SquareEventNotifiedCreateSquareChatMember;
	notifiedUpdateSquareMemberRelation:
		SquareEventNotifiedUpdateSquareMemberRelation;
	notifiedShutdownSquare: SquareEventNotifiedShutdownSquare;
	notifiedKickoutFromSquare: SquareEventNotifiedKickoutFromSquare;
	notifiedDeleteSquareChat: SquareEventNotifiedDeleteSquareChat;
	notificationJoinRequest: SquareEventNotificationJoinRequest;
	notificationJoined: SquareEventNotificationMemberUpdate;
	notificationPromoteCoadmin: SquareEventNotificationMemberUpdate;
	notificationPromoteAdmin: SquareEventNotificationMemberUpdate;
	notificationDemoteMember: SquareEventNotificationMemberUpdate;
	notificationKickedOut: SquareEventNotificationMemberUpdate;
	notificationSquareDelete: SquareEventNotificationSquareDelete;
	notificationSquareChatDelete: SquareEventNotificationSquareChatDelete;
	notificationMessage: SquareEventNotificationMessage;
	notifiedUpdateSquareChatProfileName:
		SquareEventNotifiedUpdateSquareChatProfileName;
	notifiedUpdateSquareChatProfileImage:
		SquareEventNotifiedUpdateSquareChatProfileImage;
	notifiedUpdateSquareFeatureSet: SquareEventNotifiedUpdateSquareFeatureSet;
	notifiedAddBot: SquareEventNotifiedAddBot;
	notifiedRemoveBot: SquareEventNotifiedRemoveBot;
	notifiedUpdateSquareNoteStatus: SquareEventNotifiedUpdateSquareNoteStatus;
	notifiedUpdateSquareChatAnnouncement:
		SquareEventNotifiedUpdateSquareChatAnnouncement;
	notifiedUpdateSquareChatMaxMemberCount:
		SquareEventNotifiedUpdateSquareChatMaxMemberCount;
	notificationPostAnnouncement: SquareEventNotificationPostAnnouncement;
	notificationPost: SquareEventNotificationPost;
	mutateMessage: SquareEventMutateMessage;
	notificationNewChatMember: SquareEventNotificationNewChatMember;
	notifiedUpdateReadonlyChat: SquareEventNotifiedUpdateReadonlyChat;
	notifiedUpdateMessageStatus: SquareEventNotifiedUpdateMessageStatus;
	notificationMessageReaction: SquareEventNotificationMessageReaction;
	chatPopup: SquareEventChatPopup;
	notifiedSystemMessage: SquareEventNotifiedSystemMessage;
	notifiedUpdateSquareChatFeatureSet:
		SquareEventNotifiedUpdateSquareChatFeatureSet;
	notifiedUpdateLiveTalkInfo: SquareEventNotifiedUpdateLiveTalkInfo;
	notifiedUpdateLiveTalk: SquareEventNotifiedUpdateLiveTalk;
	notificationLiveTalk: SquareEventNotificationLiveTalk;
	notificationThreadMessage: SquareEventNotificationThreadMessage;
	notificationThreadMessageReaction:
		SquareEventNotificationThreadMessageReaction;
	notifiedUpdateThread: SquareEventNotifiedUpdateThread;
	notifiedUpdateThreadStatus: SquareEventNotifiedUpdateThreadStatus;
	notifiedUpdateThreadMember: SquareEventNotifiedUpdateThreadMember;
	notifiedUpdateThreadRootMessage: SquareEventNotifiedUpdateThreadRootMessage;
	notifiedUpdateThreadRootMessageStatus:
		SquareEventNotifiedUpdateThreadRootMessageStatus;
	notifiedCreateSquareSubscription: any;
	notifiedUpdateSquareSubscription: any;
}